pyproject.toml
[tool.poetry]
name = "meta-api-course"
version = "0.1.0"
description = ""
authors = ["Cloves Paiva <cloves.paiva@extremedigital.com.br>"]
readme = "README.md"
packages = [{include = "meta_api_course"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
djangorestframework = "^3.14.0"
djangorestframework-xml = "^2.0.0"
djangorestframework-csv = "^2.1.1"
djangorestframework-yaml = "^2.0.0"
bleach = "^6.0.0"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
